money_Rotation_bot/requirements.txt
ccxt
pandas
requests
flask
python-telegram-bot